airflow_service/requirements.txt
airflow-clickhouse-plugin
apache-airflow-providers-postgres
pandas